[FIX] try to fix build bug

+++ src/ugv_else/emcl2_ros2/src/Mcl.cpp
@@ -96,12 +96,12 @@ void Mcl::sensorUpdate(double lidar_x, double lidar_y, double lidar_t, bool inv)
 	int i = 0;
 	if (!inv) {
 		for ([[maybe_unused]] auto & _ : scan.ranges_) {
-			uint16_t rep = Pose::get16bitRepresentation((double)(scan.angle_min_ + (i++) * scan.angle_increment_));
+			uint16_t rep = Pose::get16bitRepresentation(scan.angle_min_ + (i++) * scan.angle_increment_);
 			scan.directions_16bit_.push_back(rep);
 		}
 	} else {
 		for ([[maybe_unused]] auto & _ : scan.ranges_) {
-			uint16_t rep = Pose::get16bitRepresentation((double)(scan.angle_max_ - (i++) * scan.angle_increment_));
+			uint16_t rep = Pose::get16bitRepresentation(scan.angle_max_ - (i++) * scan.angle_increment_);
 			scan.directions_16bit_.push_back(rep);
 		}
 	}
